Tanya| add common methods to utils file
