+++ src/features/DisplayControls/NursingTasks/styles/AdministeredMedicationList.scss
@@ -2,7 +2,8 @@
   margin: 0;
 }
 
-.time, .status {
+.time,
+.status {
   position: relative;
   padding-left: 8px;
 }
@@ -19,9 +20,6 @@
   width: 55%;
 }
 
-.time-cell {
+.administered-time-cell {
   --day-border: 0px !important;
 }
-
-
-
