+++ src/features/DisplayControls/NursingTasks/styles/UpdateNursingTasks.scss
@@ -51,9 +51,3 @@
     margin: 0;
   }
 }
-
-.divider {
-  border-bottom: 1px solid #ccc;
-  margin: 0;
-}
-
